+++ punchpipe/control/cleaner.py
@@ -1,7 +1,9 @@
 import os
 from pathlib import Path
+from datetime import datetime, timedelta
 
-from prefect import flow, get_run_logger
+from prefect import flow, get_run_logger, task
+from prefect.cache_policies import NO_CACHE
 from sqlalchemy.orm import aliased
 
 from punchpipe.control.db import File, FileRelationship, Flow
@@ -16,6 +18,12 @@ def cleaner(pipeline_config_path: str, session=None):
     if session is None:
         session = get_database_session()
 
+    reset_revivable_flows(logger, session, pipeline_config)
+    fail_flows_stuck_as_launched(logger, session, pipeline_config)
+
+
+@task(cache_policy=NO_CACHE)
+def reset_revivable_flows(logger, session, pipeline_config):
     # Note: I thought about adding a maximum here, but this flow takes only 5 seconds to revive 10,000 L1 flows, so I
     # think we're good.
     child = aliased(File)
@@ -78,3 +86,22 @@ def cleaner(pipeline_config_path: str, session=None):
     session.commit()
     if len(unique_flows):
         logger.info(f"Processed {len(unique_flows)} revivable flows")
+
+
+@task(cache_policy=NO_CACHE)
+def fail_flows_stuck_as_launched(logger, session, pipeline_config):
+    amount_of_patience = pipeline_config['control']['cleaner'].get('fail_launched_flows_after_minutes', -1)
+    if amount_of_patience < 0:
+        return
+
+    stucks = (session.query(Flow)
+              .where(Flow.state == 'launched')
+              .where(Flow.launch_time < datetime.now() - timedelta(minutes=amount_of_patience))
+              ).all()
+
+    if len(stucks):
+        for stuck in stucks:
+            stuck.state = 'failed'
+        session.commit()
+
+        logger.info(f"Failed {len(stucks)} flows that have been in a 'launched' state for {amount_of_patience} minutes")