+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Updates replay request cleaning script in https://github.com/punch-mission/punchpipe/pull/190
 * Don't schedule for disabled flows in https://github.com/punch-mission/punchpipe/pull/203
 * Dashboard fix for file cards with multiple file types in https://github.com/punch-mission/punchpipe/pull/202
+* Cleans flows stuck in a 'launched' state in https://github.com/punch-mission/punchpipe/pull/209
 * Set FILEVRSN for files before writing in https://github.com/punch-mission/punchpipe/pull/204
 * Adds docs button to view source in https://github.com/punch-mission/punchpipe/pull/208